Finished UserProfileTest, removed second unneeded test class
